Changed web pack config. Added sub “dist” folders. renamed “modules” and “lib”
