+++ lib/info.js
