+++ modules/adobeAppMeasurement.js
@@ -1,0 +1,9 @@
+import opMan from 'lib/optionsManager'
+/**
+ * this method does nothing
+ * @private
+ */
+export default function(options){
+	var thisstuff= 'and yea';
+	opMan(options)
+}
