+++ modules/example/stuff/banner.html
