+++ modules/example/stuff/nigiri.jpg
