+++ modules/example/stuff/sidebar.html
