+++ modules/today/ads/specialpage-placements.js
@@ -1,5 +1,5 @@
-import * as utils from 'tm_tools/basicUtils';
-/* Mobile / Tablet - shared Canonical Recipe ad placements (Desktop placements can be found in desktop/ads) */
+import * as utils from 'lib/basicUtils';
+/* Desktop / Tablet - Native/Special pages - mobile form factor can be found in app/scripts/mobile/ads/specialpage-placements.js */
 
 export default function() {
   return utils.runWhenAvailable([
@@ -10,66 +10,46 @@ export default function() {
   ], function (mps, jQuery, Backbone, waypoints) {
 
 
-    var mpsDevice = mps.pagevars && mps.pagevars.cags && mps.pagevars.cags.device;
-
     var AdPlacements = Backbone.View.extend({
-      formFactor: String(mpsDevice || 'mobile'),
-      isNotLandscapeTablet: $(window).width() < 910,
+
       adTrackerRun: false,
-      boxAdHeight: 250 + 13, // includes ad label height
 
       initialize: function () {
-        this.manageContainers(); // create all ad containers
-        this.loadComponent('.revpanel-belowarticle');
+        this.manageContainers(false); // create all ad containers
         this.bindAdRenderedCallback(); // performs further adjustments based on rendered ad data
         this.bindAdsRefreshedCallback(); // if slot refreshed does not display at the responsive set size and needs cleanup, cleanup goes here
       },
 
-      manageContainers: function () {
-        var self = this,
-          skipAd = false,
-          allSlots = mps.responsiveslots[mps._loadset];
+      manageContainers: function (addWaypoints) { // renders all ad containers, even if no ads being placed at current responsive set size
+        var allSlots = mps.responsiveslots[mps._loadset];
 
-        for (var adSlotKey in allSlots) {
+        _.each(allSlots, function (slotKey, adSlotKey) {
           if (!$('#j-' + adSlotKey + '-ad').length) {
-            self.createAdContainer(adSlotKey);
+            this.createAdContainer(adSlotKey); // create each container
+            if (addWaypoints) {
+              $(window).trigger('ad-controller:add-waypoints', ['#j-' + adSlotKey + '-ad']);
+            }
           }
-        }
+        }, this);
+      },
+
+      bindMpsLoadedCallback: function () {
+        var self = this;
+        mps.initCallback = function () {
+          self.manageContainers(true); // deferred because mps._loadset updates more slowly than the containers are created, create all ad containers, attaches listener to entry change event for creating more ad containers
+        };
       },
 
       createAdContainer: function (adSlotKey) {
-        var self = this,
-          containerTemplate,
-          placement = adSlotKey,
-          typeContainsBox = adSlotKey.indexOf('box');
+        var containerTemplate,
+          placement = adSlotKey;
 
-        if (typeContainsBox !== -1) {
-          placement = 'box'; // unifying boxflex and box#s for stream placement
-        }
         containerTemplate = _.template('<div class="<%= p %>-ad"><div id="j-<%= k %>-ad" class="ad-inner<% if (ll) { %> lazy-load"<% } else { %>"<% } %> data-placement="<%= k %>"></div></div>'); // underscore template for ad containers
 
         switch (placement) { // where to render each ad by case
           case 'topbanner':
-            $('main.content-wrapper').before(containerTemplate({p: placement, k: adSlotKey, ll: false}));
+            $('#j-entry-panel').before(containerTemplate({p: placement, k: adSlotKey, ll: false}));
             break;
-          case 'box':
-            if (self.isNotLandscapeTablet && adSlotKey === 'boxflex') {
-              $('.recipe-details-container').after(containerTemplate({p: placement, k: adSlotKey, ll: true}));
-            } else if (self.formFactor === 'tablet' && !self.isNotLandscapeTablet && adSlotKey === 'box1') {
-              $('.recipe-l-panel').append(containerTemplate({p: placement, k: adSlotKey, ll: true}));
-            } else if (adSlotKey === 'boxflex') {
-              $('.food-community').after(containerTemplate({p: placement, k: adSlotKey, ll: true}));
-            } else if (self.isNotLandscapeTablet && adSlotKey === 'box1') {
-              $('.recipe-l-bottom').before(containerTemplate({p: placement, k: adSlotKey, ll: true}));
-            }
-            break;
-        }
-      },
-
-      loadComponent: function (element) {
-        var componentID = $(element).data('component-name'); // assumes data-component-id attribute on .rev-panel
-        if (componentID && $(element).children().length < 1) {
-          mps.insertComponent(mps._select(element), componentID);
         }
       },
 
@@ -109,12 +89,13 @@ export default function() {
         mps.adloadCallback = function (adData) {
           var placement = adData._mps._slot,
             slot = adData._mps.adslot,
-            id = '#' + adData._mps.adslot,
+            id = '#' + slot,
             isEmpty = adData.isEmpty,
             parentContainer = $(id).parent().parent(),
+            parentEntry = $(parentContainer).closest('.entry'),
+            parentEntryID = $(parentEntry).data('id'),
             noLabel = false, // flag to hide ad label
-            showAdChoices = false, // flag to add the extra ad choices icon and text to the ad label
-            adHeight;
+            showAdChoices = false; // flag to add the extra ad choices icon and text to the ad label
 
           self.loadedAdsTracker(placement);
 
@@ -123,22 +104,14 @@ export default function() {
           }
 
           switch (placement) {
-            case 'topbanner': // must load after rails in MPS (DART Ad Units heavier weight)
-              adHeight = adData.size[1];
-              self.topbannerWidth = adData.size[0]; // making this accessible to rails
-              noLabel = true;
-              $(id).css({'height': adHeight + 'px'});
-              break;
-            case 'boxflex':
+            case 'topbanner':
               adHeight = adData.size[1];
               showAdChoices = true;
-              if (adHeight > self.boxAdHeight) { // revised to check if is taller than row-height rather than look for a specific height
-                $(parentContainer).addClass('large');
-              } else {
-                $(parentContainer).removeClass('large');
-              }
+              $(id).css({'height': adHeight + 'px'});
+              Tdy.Events.trigger('ads:reposition-sharebar');
               break;
           }
+
           if (!noLabel) {
             $(window).trigger('ad-controller:render-label', [slot, showAdChoices, isEmpty]); // add the label via ad-controller.js
           }
@@ -146,37 +119,41 @@ export default function() {
       },
 
       bindAdsRefreshedCallback: function () {
-        var self = this;
-        mps.refreshAdsCallback = function (slotvars) {
-          var refreshedSlots = mps._adloads['R' + mps._loadset] || [], // makes sure we're getting the refreshed object of the current adloadset, normally 0/R0
-            disabledAds = mps.slotsdisabled[mps._loadset]; // disabled ad slots, not refreshed
-
+        mps.refreshAdsCallback = _.throttle(function (slotvars) { // no mechanism (yet) for discovering which ad triggered this callback, so all clean-up executed on call. Throttles so that multiple callbacks on the same resize don't each trigger it.
           var adCleanup = function (placement) {
             var containers = $('div.ad-inner[data-placement="' + placement + '"]');
+
             $.each(containers, function (index, div) {
               var slotID = $(div).find('.mps-slot').attr('id');
 
-              $(div).parent().removeClass('ad-loaded');
+              $(div).parent().removeClass('ad-loaded'); // used for items where CSS needs to be applied ONLY when ads are returned
               $(window).trigger('ad-controller:render-label', [slotID, null, true]); // removes the label via ad-controller.js for ANY disabled slot
             });
           };
 
-          $.each(refreshedSlots, function (placement, loads) { // loops through each refreshed slot to check whether or not an ad was returned
-            var empty = loads[_.size(loads) - 1].empty; // gets whether the latest ad load happened or not
-            if (empty) {
-              adCleanup(placement);
+          $.each(mps._adloads, function (key, refreshedSlots) { // loops through each refreshed set of ads to see whether any returned empty
+            if (key.substring(0, 1) === 'R') {
+              $.each(mps._adloads[key], function (placement, loads) { // loops through each refreshed slot to check whether or not an ad was returned
+                var empty = loads[_.size(loads) - 1].empty; // gets whether the latest ad load happened or not
+                if (empty) {
+                  adCleanup(placement);
+                }
+              });
             }
           });
 
-          for (var placement in disabledAds) { // disabled ads don't show up in the refreshedSlots object, this takes care of them
-            adCleanup(disabledAds[placement]);
-          }
-        };
+          $.each(mps.slotsdisabled, function (key, disabledSlots) { // loops through each disabled set of ads
+            var disabledSet = mps.slotsdisabled[key];
+            $.each(disabledSet, function (placement) { // disabled ads don't show up in the refreshedSlots object, this takes care of them
+              adCleanup(disabledSet[placement]);
+            });
+          });
+
+        }, 500);
       }
 
     });
 
     return AdPlacements;
   });
-
 };
