+++ package.json
@@ -1,19 +1,25 @@
 {
-  "name": "hubtm",
-  "version": "1.0.0",
-  "description": "github tag management",
-  "main": "index.js",
+  "name": "hubTm",
+  "version": "0.1",
+  "description": "script bundler and deployment tool",
+  "main": "bundle.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "packTm": "webpack --config webpack.config.js"
+    "test": "echo \"Error: no test specified\" && exit 1"
   },
   "author": "Jeremy Fletcher",
+  "contributors": [
+    {
+      "name": "Todd Wallar",
+      "email": "todd.wallar@nbcuni.com"
+    }
+  ],
   "license": "ISC",
   "dependencies": {
     "css-loader": "^0.26.1",
     "file-loader": "^0.9.0",
     "html-loader": "^0.4.4",
     "style-loader": "^0.13.1",
+    "url-loader": "^0.5.7",
     "webpack": "^2.1.0-beta.27"
   },
   "repository": {
