--- tm_modules/adobeAppMeasurement.js
@@ -1,9 +1,0 @@
-import opMan from 'tm_tools/optionsManager'
-/**
- * this method does nothing
- * @private
- */
-export default function(options){
-	var thisstuff= 'and yea';
-	opMan(options)
-}
